+++ pkg/engine/mutate/patch/patchesUtils_test.go
@@ -1,7 +1,6 @@
 package patch
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/go-logr/logr"
@@ -131,89 +130,6 @@ var overlayBytes = []byte(`
   }
 }
 `)
-
-var expectBytes = []byte(`{"apiVersion": "apps/v1","kind": "Deployment","metadata": {"name": "wordpress","labels": {"app": "wordpress"}},"spec": {"selector": {"matchLabels": {"app": "wordpress"}},"strategy": {"type": "Recreate"},"template": {"metadata": {"labels": {"app": "wordpress"}},"spec": {"containers": [{"name": "nginx","image": "nginx"},{"image": "wordpress:4.8-apache","name": "wordpress","ports": [{"containerPort": 80,"name": "wordpress"}],"volumeMounts": [{"name": "wordpress-persistent-storage","mountPath": "/var/www/html"}],"env": [{"name": "WORDPRESS_DB_HOST","value": "$(MYSQL_SERVICE)"},{"name": "WORDPRESS_DB_PASSWORD","valueFrom": {"secretKeyRef": {"name": "mysql-pass","key": "password"}}}]}],"volumes": [{"name": "wordpress-persistent-storage"}],"initContainers": [{"name": "init-command","image": "debian","command": ["echo $(WORDPRESS_SERVICE)","echo $(MYSQL_SERVICE)"]}]}}}}`)
-
-func Test_ignorePath(t *testing.T) {
-	tests := []struct {
-		path   string
-		ignore bool
-	}{
-		{
-			path:   "/",
-			ignore: false,
-		},
-		{
-			path:   "/metadata",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/name",
-			ignore: false,
-		},
-		{
-			path:   "spec/template/metadata/name",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/namespace",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/annotations",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/labels",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/ownerReferences",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/finalizers",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/generateName",
-			ignore: false,
-		},
-		{
-			path:   "/metadata/creationTimestamp",
-			ignore: true,
-		},
-		{
-			path:   "spec/template/metadata/creationTimestamp",
-			ignore: true,
-		},
-		{
-			path:   "/metadata/resourceVersion",
-			ignore: true,
-		},
-		{
-			path:   "/status",
-			ignore: false,
-		},
-		{
-			path:   "/spec",
-			ignore: false,
-		},
-		{
-			path:   "/kind",
-			ignore: false,
-		},
-		{
-			path:   "/spec/triggers/0/metadata/serverAddress",
-			ignore: false,
-		},
-	}
-
-	for _, test := range tests {
-		res := ignorePatch(test.path)
-		assert.Equal(t, test.ignore, res, fmt.Sprintf("test fails at %s", test.path))
-	}
-}
 
 func Test_GeneratePatches_sortRemovalPatches(t *testing.T) {
 	base := []byte(`{"apiVersion": "apps/v1","kind": "Deployment","metadata": {"name": "nginx-deployment","labels": {"app": "nginx"}},"spec": {"selector": {"matchLabels": {"app": "nginx"}},"replicas": 1,"template": {"metadata": {"labels": {"app": "nginx"}},"spec": {"containers": [{"name": "nginx","image": "nginx:1.14.2","ports": [{"containerPort": 80}]}],"tolerations": [{"effect": "NoExecute","key": "node.kubernetes.io/not-ready","operator": "Exists","tolerationSeconds": 300},{"effect": "NoExecute","key": "node.kubernetes.io/unreachable","operator": "Exists","tolerationSeconds": 300}]}}}}`)
